mshack2019/requirements.txt
flask-restplus
flask-cors
werkzeug
blinker==1.4
azure-cognitiveservices-vision-computervision
pillow